% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{longrun}
\alias{longrun}
\title{linear time longest run search: find the longest run of values in the
ordered vector within the specified limit}
\usage{
longrun(v, limit, group = NULL)
}
\arguments{
\item{v}{a vector with entries ordered in increasing order}

\item{limit}{distance interval}

\item{group}{optional vector with entries -1 and 1 corresponding to the group membership of each element in \code{v} (two groups)}
}
\value{
run length
}
\description{
linear time longest run search: find the longest run of values in the
ordered vector within the specified limit
}
\keyword{internal}

