% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{longrun}
\alias{longrun}
\title{linear time longest run search: find the longest run of values in the
ordered vector within the specified limit}
\usage{
longrun(v, limit)
}
\arguments{
\item{v}{a vector with entries ordered in increasing order}

\item{limit}{distance interval}
}
\value{
run length
}
\description{
linear time longest run search: find the longest run of values in the
ordered vector within the specified limit
}
\keyword{internal}

