% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdist.R
\name{tdist}
\alias{tdist}
\title{Thresholded distances between columns of a matrix.}
\usage{
tdist(A, t, p = 10, filter = c("distributed", "local"),
  method = c("euclidean", "manhattan"), dry_run = FALSE, restart, ...)
}
\arguments{
\item{A}{an m by n real-valued dense or sparse matrix}

\item{t}{a threshold distance value, if missing an estimate derived from a 1-d SVD projection will be used}

\item{p}{projected subspace dimension}

\item{filter}{"local" filters candidate set sequentially,
"distributed" computes thresholded correlations in a parallel code section which can be
faster but requires that the data matrix is available (see notes).}

\item{method}{the distance measure to be used, one of
         "euclidean", or "manhattan".
Any unambiguous substring can be given.}

\item{dry_run}{set \code{TRUE} to return statistics and truncated SVD for tuning
\code{p} (see notes)}

\item{restart}{either output from a previous run of \code{tdist} with \code{dry_run=TRUE},
or direct output from from \code{\link{irlba}} used to restart the \code{irlba}
algorithm when tuning \code{p} (see notes)}

\item{...}{additional arguments passed to \code{\link{irlba}}}
}
\value{
A list with elements:
\enumerate{
  \item \code{indices} A three-column matrix. The  first two columns contain
        indices of vectors meeting the distance threshold \code{t},
        the third column contains the corresponding distance value.
  \item \code{longest_run} The largest number of successive entries in the
    ordered first singular vector within a projected distance defined by the
    correlation threshold.
  \item \code{n} The total number of _possible_ vectors that meet
    the correlation threshold identified by the algorithm.
  \item \code{total_time} Total run time.
}
}
\description{
Unlike the \code{\link{dist}} function which computes
distances between matrix _rows_, \code{tdist} computes distances
exceeding a threshold between matrix _columns_.
}
\details{
Increase p to cut down the total number of candidate pairs evaluated,
at the expense of costlier truncated SVDs.

This function doesn't work as well generally yet as the \code{\link{tcor}}
companion function in this package.
}

